+++ toggle_speech_optimization.py
@@ -15,6 +15,7 @@ class Toggle_Speech_Optimization():
         self.roles_blacklist = []
         self.on_message = [self.dronemode]
         self.on_ready = []
+        self.help_content = {'name': 'optimize', 'value': 'optimize drone speech patterns; this command can only be used by the Hive Mxtress'}
 
     async def dronemode(self, message: discord.Message):