+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - the AI now requires Python version 3.8 or higher
 - DroneMode renamed to Speech Optimization
+- Added orders reporting module
 
 ## [1.1.0.5]
 ### Fixed