+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+## Added
+- added a `help` command that tells users which command they can use in the current channel
+
 ### Changed
 - the AI now requires Python version 3.8 or higher
 - DroneMode renamed to Speech Optimization